Replace JVM dependencies: add kotlin.jvm imports, Math→kotlin.math, Double.valueOf removal
- Add explicit kotlin.jvm.JvmField/JvmStatic/JvmOverloads imports to 137 files
  (required for commonMain compilation where these aren't auto-imported)
- Replace java.lang.Math with kotlin.math equivalents in 22 files
- Replace java.lang.Double.valueOf() with Kotlin idioms in 12 files
- All 710+ JVM tests pass

+++ commcare-core/src/main/java/org/javarosa/xpath/expr/XPathCountFunc.kt
@@ -17,7 +17,7 @@ open class XPathCountFunc : XPathFuncExpr {
 
     override fun evalBody(model: DataInstance<*>?, evalContext: EvaluationContext, evaluatedArgs: Array<Any?>): Any {
         if (evaluatedArgs[0] is XPathNodeset) {
-            return java.lang.Double.valueOf((evaluatedArgs[0] as XPathNodeset).size().toDouble())
+            return (evaluatedArgs[0] as XPathNodeset).size().toDouble()
         } else {
             throw XPathTypeMismatchException("uses an invalid reference inside a count function")
         }
